src/utils/getBat.spec.ts
import { getBat } from './getBat.js'

describe('getBat', () => {
	const metadata = {
		$lastUpdated: '2023-07-07T12:11:03.0324459Z',
		lwm2m: {
			'3': {
				'0': {
					'0': {
						$lastUpdated: '2023-07-07T12:11:03.0324459Z',
						value: {
							$lastUpdated: '2023-07-07T12:11:03.0324459Z',
						},
					},
					'3': {
						$lastUpdated: '2023-07-07T12:11:03.0324459Z',
						value: {
							$lastUpdated: '2023-07-07T12:11:03.0324459Z',
						},
					},
					'7': {
						$lastUpdated: '2023-08-03T12:11:03.0324459Z',
						value: {
							$lastUpdated: '2023-08-03T12:11:03.0324459Z',
						},
					},
					'13': {
						$lastUpdated: '2023-07-07T12:11:03.0324459Z',
						value: {
							$lastUpdated: '2023-07-07T12:11:03.0324459Z',
						},
					},
					$lastUpdated: '2023-07-07T12:11:03.0324459Z',
				},
				$lastUpdated: '2023-07-07T12:11:03.0324459Z',
			},
			$lastUpdated: '2023-07-07T12:11:03.0324459Z',
		},
	}

	it(`should create the 'bat' object expected by nRF Asset Tracker`, () => {
		const device = {
			'0': 'Nordic Semiconductor ASA',
			'1': 'Thingy:91',
			'2': '351358815340515',
			'3': '22.8.1+0',
			'7': [2754],
			'11': [0],
			'13': 1675874731,
			'16': 'UQ',
			'19': '3.2.1',
		}
		const bat = getBat(device, metadata) as { result: unknown }
		expect(bat.result).toStrictEqual({
			v: 2754,
			ts: 1675874731000,
		})
	})

	/**
	 * @see adr/007-timestamp-hierarchy.md
	 */
	it('should follow Timestamp Hierarchy in case timestamp is not found from Device object', () => {
		const device = {
			'0': 'Nordic Semiconductor ASA',
			'1': 'Thingy:91',
			'2': '351358815340515',
			'3': '22.8.1+0',
			'11': [0],
			'7': [80],
			// '13': 1675874731, // timestamp from Device object
			'16': 'UQ',
			'19': '3.2.1',
		}

		const battery = getBat(device, metadata) as { result: unknown }
		expect(battery.result).toMatchObject({
			v: 80,
			ts: 1688731863032,
		})
	})

	it(`should return error if Device object is undefined`, () => {
		const result = getBat(undefined, metadata) as { error: Error }
		expect(result.error).not.toBe(undefined)
	})

	it(`should return error if conversion from Device object to 'bat' object went wrong`, () => {
		const device = {
			'0': 'Nordic Semiconductor ASA',
			'1': 'Thingy:91',
			'2': '351358815340515',
			'3': '22.8.1+0',
			// '7':  [2754], required value missed
			'11': [0],
			'13': 1675874731,
			'16': 'UQ',
			'19': '3.2.1',
		}
		const bat = getBat(device, metadata) as { error: Error }
		expect(bat.error).not.toBe(undefined)
	})
})
